+++ src/core/managers/internal/PendingTransactionsStateManager/types/pendingTransactions.types.ts
@@ -1,4 +1,4 @@
-// types here need to be synced with the types in sdk-dapp-core-ui
+// types here need to be synced with the types in sdk-dapp-ui
 export enum PendingTransactionsEventsEnum {
   CLOSE = 'CLOSE_PENDING_TRANSACTIONS',
   DATA_UPDATE = 'DATA_UPDATE_PENDING_TRANSACTIONS'