+++ src/core/managers/internal/SignTransactionsStateManager/SignTransactionsStateManager.ts
@@ -1,7 +1,7 @@
 import { EMPTY_PPU } from 'constants/placeholders.constants';
 import { UITagsEnum } from 'constants/UITags.enum';
 import { Transaction } from 'lib/sdkCore';
-import { MvxSignTransactionsPanel } from 'lib/sdkDappCoreUi';
+import { MvxSignTransactionsPanel } from 'lib/sdkDappUi';
 import { NftEnumType } from 'types/tokens.types';
 import {
   FungibleTransactionType,